requirements.txt
transformers==4.36.0
datasets==2.14.0
accelerate==0.24.0
evaluate==0.4.0
rouge-score==0.1.2
nltk==3.8.1
faiss-cpu==1.7.4
bert-score==0.3.13
pandas==2.0.3
sacrebleu==2.3.1
torch>=2.0.0
langchain==0.0.350
streamlit==1.32.2
gradio==4.8.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0
openpyxl>=3.1.0
